New Endpoints and pg migration for production

+++ data/seeds/00_cleanup.js
@@ -1,5 +1,7 @@
-exports.seed = async function(knex) {
-	await knex("moods_by_date").truncate()
-	await knex("sleep_entries").truncate()
-	await knex("users").truncate()
+exports.seed = async function (knex) {
+	if (process.env.NODE_ENV != "production") {
+		await knex("moods_by_date").truncate()
+		await knex("sleep_entries").truncate()
+		await knex("users").truncate()
+	}
 }
